Fixing Conflicting Issuers In Database
This takes all the securities in our database and looks for cases where one s_InvIssuer (raw NMFP name for a security) has been classified as two different issuers. It selects the most frequent issuer classified for that security. This is used for future issuer classification.
i.	Output:
1.	fix_issuer_cusip_diffs.pkl 
a.	This file will be used in old_issuers.py